Now seeding 14 mods into the database Added seeding configuration option to seed or not seed users

+++ mod-picker/db/seeds.rb
@@ -1,4 +1,10 @@
 #==================================================
+# CONFIGURATION OPTIONS
+#==================================================
+
+bSeedUsers = false
+
+#==================================================
 # CLEAR TABLES
 #==================================================
 
@@ -12,10 +18,12 @@ Mod.delete_all
 connection.execute("DELETE FROM categories WHERE parent_id IS NOT NULL;")
 connection.execute("DELETE FROM categories WHERE parent_id IS NULL;")
 # clear users and associated tables
-UserBio.delete_all
-UserReputation.delete_all
-UserSetting.delete_all
-User.delete_all
+if (bSeedUsers)
+  UserBio.delete_all
+  UserReputation.delete_all
+  UserSetting.delete_all
+  User.delete_all
+end
 # clear games
 Game.delete_all
 
@@ -335,44 +343,46 @@ Category.create(
 #==================================================
 require 'securerandom'
 
-# create an admin user
-User.create!(
-    username: "admin",
-    user_level: "admin",
-    title: "God",
-    avatar: "",
-    joined: Time.now.to_date,
-    email: "admin@mail.com",
-    password: SecureRandom.urlsafe_base64,
-    encrypted_password: SecureRandom.urlsafe_base64,
-    reset_password_token: "foobarpw",
-    sign_in_count: 1,
-    current_sign_in_at: Time.now.to_date,
-    last_sign_in_at: Time.now.to_date,
-    current_sign_in_ip: Faker::Internet.public_ip_v4_address,
-    last_sign_in_ip: Faker::Internet.public_ip_v4_address
-)
-
-# generates random date between year 2000 and now.
-def time_rand from = Time.new(2000), to = Time.now
-  Time.at(from + rand * (to.to_f - from.to_f)).to_date
-end
-
-# create 99 random users
-99.times do |n|
-  name = Faker::Internet.user_name
+if (bSeedUsers)
+  # create an admin user
   User.create!(
-      username: "#{name}#{n}",
-      joined: time_rand,
-      email: Faker::Internet.email(name),
+      username: "admin",
+      user_level: "admin",
+      title: "God",
+      avatar: "",
+      joined: Time.now.to_date,
+      email: "admin@mail.com",
       password: SecureRandom.urlsafe_base64,
       encrypted_password: SecureRandom.urlsafe_base64,
-      reset_password_token: Faker::Internet.password,
-      sign_in_count: Random.rand(100).to_i + 1,
-      current_sign_in_ip: Faker::Internet.public_ip_v4_address,
+      reset_password_token: "foobarpw",
+      sign_in_count: 1,
+      current_sign_in_at: Time.now.to_date,
       last_sign_in_at: Time.now.to_date,
+      current_sign_in_ip: Faker::Internet.public_ip_v4_address,
       last_sign_in_ip: Faker::Internet.public_ip_v4_address
   )
+
+  # generates random date between year 2000 and now.
+  def time_rand from = Time.new(2000), to = Time.now
+    Time.at(from + rand * (to.to_f - from.to_f)).to_date
+  end
+
+  # create 99 random users
+  99.times do |n|
+    name = Faker::Internet.user_name
+    User.create!(
+        username: "#{name}#{n}",
+        joined: time_rand,
+        email: Faker::Internet.email(name),
+        password: SecureRandom.urlsafe_base64,
+        encrypted_password: SecureRandom.urlsafe_base64,
+        reset_password_token: Faker::Internet.password,
+        sign_in_count: Random.rand(100).to_i + 1,
+        current_sign_in_ip: Faker::Internet.public_ip_v4_address,
+        last_sign_in_at: Time.now.to_date,
+        last_sign_in_ip: Faker::Internet.public_ip_v4_address
+    )
+  end
 end
 
 
@@ -381,7 +391,7 @@ end
 #==================================================
 
 # Helper vars
-nexusDateFormat = "%k:%M, %d %b %Y"
+nexusDateFormat = "%d/%m/%Y - %I:%M%p"
 
 # Top recently endorsed mods
 Mod.create(
@@ -392,15 +402,15 @@ Mod.create(
 )
 
 NexusInfo.create(
-    mod_id: Mod.last.id,
+    mod_id: 3863,
     uploaded_by: "schlangster",
     authors: "SkyUI Team",
-    date_released: DateTime.strptime(" 0:24, 17 Dec 2011", nexusDateFormat),
-    date_updated: DateTime.strptime("15:38, 24 Aug 2015", nexusDateFormat),
-    endorsements: 401363,
-    total_downloads: 11193037,
-    unique_downloads: 5677583,
-    views: 26147936,
+    date_released: DateTime.strptime("17/12/2011 - 12:24AM", nexusDateFormat),
+    date_updated: DateTime.strptime("24/08/2015 - 03:38PM", nexusDateFormat),
+    endorsements: 401489,
+    total_downloads: 11195070,
+    unique_downloads: 5678653,
+    views: 26151968,
     posts_count: 17023,
     videos_count: 23,
     images_count: 23,
@@ -408,3 +418,321 @@ NexusInfo.create(
     articles_count: 1,
     nexus_category: 42
 )
+
+Mod.create(
+    name: "Immersive Armors",
+    primary_category: Category.where(name: "Items - Armor, Clothing, & Accessories").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 19733,
+    uploaded_by: "hothtrooper44",
+    authors: "Hothtrooper44",
+    date_released: DateTime.strptime("01/07/2012 - 07:23PM", nexusDateFormat),
+    date_updated: DateTime.strptime("11/01/2016 - 01:51AM", nexusDateFormat),
+    endorsements: 206379,
+    total_downloads: 6391803,
+    unique_downloads: 2920317,
+    views: 16140841,
+    posts_count: 15626,
+    videos_count: 16,
+    images_count: 453,
+    files_count: 25,
+    articles_count: 3,
+    nexus_category: 54
+)
+
+Mod.create(
+    name: "Skyrim HD - 2K Textures",
+    primary_category: Category.where(name: "Audiovisual - Models & Textures").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 607,
+    uploaded_by: "NebuLa1",
+    authors: "NebuLa from AHBmods",
+    date_released: DateTime.strptime("19/11/2011 - 01:03AM", nexusDateFormat),
+    date_updated: DateTime.strptime("09/01/2016 - 04:57PM", nexusDateFormat),
+    endorsements: 156555,
+    total_downloads: 17682800,
+    unique_downloads: 8505445,
+    views: 20440708,
+    posts_count: 8472,
+    videos_count: 26,
+    images_count: 161,
+    files_count: 23,
+    articles_count: 0,
+    nexus_category: 29
+)
+
+Mod.create(
+    name: "RaceMenu",
+    primary_category: Category.where(name: "Gameplay - User Interface").first.id,
+    secondary_category: Category.where(name: "Resources - Frameworks").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 29624,
+    uploaded_by: "expired6978",
+    authors: "Expired",
+    date_released: DateTime.strptime("08/01/2013 - 01:10AM", nexusDateFormat),
+    date_updated: DateTime.strptime("18/01/2016 - 10:37AM", nexusDateFormat),
+    endorsements: 147455,
+    total_downloads: 4160279,
+    unique_downloads: 2874155,
+    views: 9600050,
+    posts_count: 9998,
+    videos_count: 8,
+    images_count: 258,
+    files_count: 104,
+    articles_count: 0,
+    nexus_category: 42
+)
+
+Mod.create(
+    name: "Unofficial Skyrim Legendary Edition Patch",
+    primary_category: Category.where(name: "Fixes").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 71214,
+    uploaded_by: "Arthmoor",
+    authors: "Unofficial Patch Project Team ",
+    date_released: DateTime.strptime("07/11/2015 - 08:41PM", nexusDateFormat),
+    date_updated: DateTime.strptime("11/01/2016 - 07:37PM", nexusDateFormat),
+    endorsements: 13221,
+    total_downloads: 249042,
+    unique_downloads: 168964,
+    views: 681548,
+    posts_count: 1604,
+    videos_count: 1,
+    images_count: 15,
+    files_count: 1,
+    articles_count: 0,
+    nexus_category: 84
+)
+
+Mod.create(
+    name: "Mod Organizer",
+    primary_category: Category.where(name: "Utilities - Tools").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 1334,
+    uploaded_by: "Tannin42",
+    authors: "Tannin",
+    date_released: DateTime.strptime("24/11/2011 - 03:30PM", nexusDateFormat),
+    date_updated: DateTime.strptime("01/12/2015 - 09:23PM", nexusDateFormat),
+    endorsements: 87747,
+    total_downloads: 2062125,
+    unique_downloads: 1496783,
+    views: 4795316,
+    posts_count: 12997,
+    videos_count: 35,
+    images_count: 4,
+    files_count: 27,
+    articles_count: 0,
+    nexus_category: 39
+)
+
+Mod.create(
+    name: "Skyrim Flora Overhaul",
+    primary_category: Category.where(name: "Audiovisual - Models & Textures").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 141,
+    uploaded_by: "vurt",
+    authors: "vurt",
+    date_released: DateTime.strptime("13/11/2011 - 10:36PM", nexusDateFormat),
+    date_updated: DateTime.strptime("16/01/2016 - 10:15PM", nexusDateFormat),
+    endorsements: 98473,
+    total_downloads: 5201828,
+    unique_downloads: 3329262,
+    views: 10208445,
+    posts_count: 10565,
+    videos_count: 47,
+    images_count: 741,
+    files_count: 35,
+    articles_count: 0,
+    nexus_category: 29
+)
+
+Mod.create(
+    name: "TES5Edit",
+    primary_category: Category.where(name: "Utilities - Tools").first.id,
+    secondary_category: Category.where(name: "Resources - Frameworks").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 25859,
+    uploaded_by: "Sharlikran",
+    authors: "ElminsterAU",
+    date_released: DateTime.strptime("22/10/2012 - 06:22AM", nexusDateFormat),
+    date_updated: DateTime.strptime("10/11/2015 - 07:43AM", nexusDateFormat),
+    endorsements: 79734,
+    total_downloads: 1586957,
+    unique_downloads: 1154565,
+    views: 4577626,
+    posts_count: 0,
+    videos_count: 14,
+    images_count: 4,
+    files_count: 6,
+    articles_count: 0,
+    nexus_category: 39
+)
+
+# Top mods in the last 59 days
+Mod.create(
+    name: "Merge Plugins",
+    primary_category: Category.where(name: "Utilities - Tools").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 69905,
+    uploaded_by: "matortheeternal",
+    authors: "Mator",
+    date_released: DateTime.strptime("24/12/2015 - 01:56AM", nexusDateFormat),
+    date_updated: DateTime.strptime("05/01/2016 - 09:16AM", nexusDateFormat),
+    endorsements: 1024,
+    total_downloads: 11444,
+    unique_downloads: 10295,
+    views: 84413,
+    posts_count: 661,
+    videos_count: 0,
+    images_count: 15,
+    files_count: 3,
+    articles_count: 0,
+    nexus_category: 39
+)
+
+Mod.create(
+    name: "The Lily - Armour Mashup",
+    primary_category: Category.where(name: "Items - Armor, Clothing, & Accessories").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 71843,
+    uploaded_by: "pottoply",
+    authors: "pottoply",
+    date_released: DateTime.strptime("08/12/2015 - 09:42PM", nexusDateFormat),
+    date_updated: DateTime.strptime("01/01/2016 - 08:21PM", nexusDateFormat),
+    endorsements: 881,
+    total_downloads: 22009,
+    unique_downloads: 18444,
+    views: 119147,
+    posts_count: 206,
+    videos_count: 5,
+    images_count: 45,
+    files_count: 15,
+    articles_count: 0,
+    nexus_category: 54
+)
+
+Mod.create(
+    name: "THE PEOPLE OF SKYRIM ULTIMATE EDITION",
+    primary_category: Category.where(name: "Gameplay - Immersion & Role-playing").first.id,
+    secondary_category: Category.where(name: "New Characters").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 72449,
+    uploaded_by: "nesbit098",
+    authors: "Nesbit",
+    date_released: DateTime.strptime("05/01/2016 - 10:31AM", nexusDateFormat),
+    date_updated: DateTime.strptime("19/01/2016 - 03:32AM", nexusDateFormat),
+    endorsements: 740,
+    total_downloads: 48840,
+    unique_downloads: 31636,
+    views: 169100,
+    posts_count: 1215,
+    videos_count: 0,
+    images_count: 18,
+    files_count: 29,
+    articles_count: 3,
+    nexus_category: 78
+)
+
+Mod.create(
+    name: "Skaal You Need - Skaal house and follower",
+    primary_category: Category.where(name: "Locations - New Player Homes").first.id,
+    secondary_category: Category.where(name: "New Characters - Allies").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 72005,
+    uploaded_by: "Elianora",
+    authors: "Elianora",
+    date_released: DateTime.strptime("17/12/2015 - 09:23PM", nexusDateFormat),
+    date_updated: DateTime.strptime("29/12/2015 - 08:16PM", nexusDateFormat),
+    endorsements: 705,
+    total_downloads: 11941,
+    unique_downloads: 9776,
+    views: 94548,
+    posts_count: 240,
+    videos_count: 2,
+    images_count: 35,
+    files_count: 4,
+    articles_count: 0,
+    nexus_category: 67
+)
+
+Mod.create(
+    name: "SC - Hairstyles",
+    primary_category: Category.where(name: "Character Appearance - Face Parts").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 71561,
+    uploaded_by: "ShinglesCat",
+    authors: "ShinglesCat",
+    date_released: DateTime.strptime("25/11/2015 - 09:21AM", nexusDateFormat),
+    date_updated: DateTime.strptime("02/01/2016 - 06:11PM", nexusDateFormat),
+    endorsements: 632,
+    total_downloads: 10011,
+    unique_downloads: 8458,
+    views: 89746,
+    posts_count: 98,
+    videos_count: 3,
+    images_count: 61,
+    files_count: 2,
+    articles_count: 0,
+    nexus_category: 26
+)
+
+Mod.create(
+    name: "Real Names",
+    primary_category: Category.where(name: "Gameplay - Immersion & Role-playing").first.id,
+    game_id: gameSkyrim.id
+)
+
+NexusInfo.create(
+    mod_id: 71464,
+    uploaded_by: "nellshini",
+    authors: "Jaxonz and Nellshini",
+    date_released: DateTime.strptime("23/11/2015 - 02:11AM", nexusDateFormat),
+    date_updated: DateTime.strptime("16/01/2016 - 08:36PM", nexusDateFormat),
+    endorsements: 628,
+    total_downloads: 10530,
+    unique_downloads: 9121,
+    views: 45741,
+    posts_count: 163,
+    videos_count: 0,
+    images_count: 6,
+    files_count: 3,
+    articles_count: 0,
+    nexus_category: 78
+)
+
